13/03/2019

+++ Assets/Askowl/Decoupler/Scripts/Template/Wizard.cs
@@ -23,23 +23,25 @@ namespace Decoupler {
     protected override void Clear() => newTemplateServiceName = "";
 
     protected override void Create() {
-      assetType = "/*-assetType-*/";
+      assetType   = "/*-assetType-*/";
+      destination = "/*-destination-*/";
       CreateAssets("cs");
     }
     protected override bool ProcessAllFiles(string textAssetTypes) {
       var serviceFor = $"{destination}/_Template_ServiceFor";
-      var fileName   = $"{serviceFor}{newTemplateServiceName}.cs";
+      if (File.Exists($"{serviceFor}.cs")) throw new Exception($"'{serviceFor}.cs already exists. Try another name.");
+      var fileName = $"{serviceFor}{newTemplateServiceName}.cs";
       File.Copy($"{serviceFor}.cs", fileName);
       ProcessFiles("cs", fileName);
       return true;
     }
     /// <a href=""></a> //#TBD#//
-    protected override string GetDestinationPath() => "/*-destination-*/";
+    protected override string GetDestinationPath() => null;
 
     protected override string FillTemplate(Template template, string text) =>
       template.From(text)
               .Substitute("_ConcreteService_", newTemplateServiceName)
-              .And("/\\*++", "").And("++\\*/", "").Result();
+              .And(@"/\*\+\+", "").And(@"\+\+\*/", "").Result();
 
     protected override void OnScriptReload() =>
       CreateAssetDictionary((newTemplateServiceName, Type.GetType($"Decoupler.Services.{newTemplateServiceName}")));
